finished the project, moved the frontend to this repo, prolly need to do something about submodule but yea its done

+++ api/views.py
@@ -43,51 +43,53 @@ def getRoutes(request):
     ]
     return Response(routes)
 
-@api_view(['GET'])
+@api_view(['GET', 'POST'])
 def getNotes(request):
-    # for all notes
-    notes = Note.objects.all().order_by('-updated')
-    serializer = NoteSerializer(notes, many=True)
     
-    return Response(serializer.data)
+    if request.method == 'GET':
+        # for all notes
+        notes = Note.objects.all().order_by('-updated')
+        serializer = NoteSerializer(notes, many=True)
+        
+        return Response(serializer.data)
+
+    # for creating a note
+    if request.method == 'POST':
+        data = request.data
+        note = Note.objects.create(body=data['body'])
+        serializer = NoteSerializer(note, many = False)
+        return Response(serializer.data)
 
 
-@api_view(['GET'])
+@api_view(['GET', 'PUT', 'DELETE'])
 def getNote(request, pk):
-    # for single note so may pk
-    notes = Note.objects.get(id=pk)
-    serializer = NoteSerializer(notes, many=False)
     
-    return Response(serializer.data)
+    if request.method == 'GET':
+        # for single note so may pk
+        notes = Note.objects.get(id=pk)
+        serializer = NoteSerializer(notes, many=False)
+        
+        return Response(serializer.data)
+    
+    # for updating a note
+    if request.method == 'PUT':
+        data = request.data
+        note = Note.objects.get(id=pk)
+        serializer = NoteSerializer(instance=note, data=data)
+        
+        if serializer.is_valid():
+            serializer.save()
+        
+        return Response(serializer.data)
+    
+    
+    #for deleting a note
+    if request.method == 'DELETE':
+        note = Note.objects.get(id=pk)
+        note.delete()
+        return Response('Note deleted')
+    
+    
+    
 
-@api_view(['PUT']) #put since update
-def updateNote(request,pk):
-    
-    data = request.data
-    note = Note.objects.get(id=pk)
-    serializer = NoteSerializer(instance=note, data=data)
-    
-    if serializer.is_valid():
-        serializer.save()
-    
-    return Response(serializer.data)
 
-
-@api_view(['DELETE'])
-def deleteNote(request, pk):
-    note = Note.objects.get(id=pk)
-    note.delete()
-    return Response('Note deleted')
-
-
-@api_view(['POST'])
-def createNote(request):
-    data = request.data
-    
-    note = Note.objects.create(
-        body=data['body']
-    )
-    
-    serializer = NoteSerializer(note, many = False)
-    
-    return Response(serializer.data)
